pyproject.toml
[project]
name = "ragmap"
version = "0.1.0"
description = "Recon tool for RAG pipelines"
requires-python = ">=3.12"
dependencies = [
    "typer",
    "httpx",
    "pydantic",
    "rich",
    "pyyaml",
    "tiktoken>=0.7.0",
]

[project.scripts]
ragmap = "ragmap.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ragmap"]

[dependency-groups]
dev = [
    "ruff>=0.15.10",
]
